requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
ansiwrap==0.8.4
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
bitstring==3.1.9
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
colorama==0.4.4
commonmark==0.9.1
cryptography==39.0.1
emoji==1.6.3
frozenlist==1.3.3
idna==3.3
importlib-metadata==6.0.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jrnl==3.3
keyring==23.13.1
MarkupSafe==2.1.1
more-itertools==9.0.0
multidict==6.0.4
openai==0.26.4
parsedatetime==2.6
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
pyxdg==0.28
PyYAML==6.0
requests==2.27.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
textwrap3==0.9.2
toncli==0.0.38
tornado==6.1
tqdm==4.64.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarl==1.8.2
zipp==3.13.0
